Move export json endpoint to export controller
