+++ src/api/Controllers/BoreholeController.cs
@@ -1,15 +1,10 @@
 ﻿using BDMS.Authentication;
 using BDMS.Models;
-using CsvHelper;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using NetTopologySuite.Geometries;
 using System.ComponentModel.DataAnnotations;
-using System.Globalization;
-using System.Text;
-using System.Text.Json;
-using System.Text.Json.Serialization;
 
 namespace BDMS.Controllers;
 
@@ -87,7 +82,7 @@ public class BoreholeController : BoreholeControllerBase<Borehole>
         pageSize = Math.Min(MaxPageSize, Math.Max(1, pageSize));
 
         var skip = (pageNumber - 1) * pageSize;
-        var query = GetBoreholesWithIncludes().AsNoTracking();
+        var query = Context.Boreholes.GetAllWithIncludes().AsNoTracking();
 
         if (ids != null && ids.Any())
         {
@@ -109,7 +104,7 @@ public class BoreholeController : BoreholeControllerBase<Borehole>
     [Authorize(Policy = PolicyNames.Viewer)]
     public async Task<ActionResult<Borehole>> GetByIdAsync(int id)
     {
-        var borehole = await GetBoreholesWithIncludes()
+        var borehole = await Context.Boreholes.GetAllWithIncludes()
             .AsNoTracking()
             .SingleOrDefaultAsync(l => l.Id == id)
             .ConfigureAwait(false);
@@ -120,92 +115,6 @@ public class BoreholeController : BoreholeControllerBase<Borehole>
         }
 
         return Ok(borehole);
-    }
-
-    /// <summary>
-    /// Asynchronously gets all <see cref="Borehole"/> records filtered by ids. Additional data is included in the response.
-    /// </summary>
-    /// <param name="ids">The required list of borehole ids to filter by.</param>
-    [HttpGet("json")]
-    [Authorize(Policy = PolicyNames.Viewer)]
-    public async Task<ActionResult> ExportJsonAsync([FromQuery][MaxLength(MaxPageSize)] IEnumerable<int> ids)
-    {
-        if (ids == null || !ids.Any()) return BadRequest("The list of IDs must not be empty.");
-
-        var boreholes = await GetBoreholesWithIncludes().AsNoTracking().Where(borehole => ids.Contains(borehole.Id)).ToListAsync().ConfigureAwait(false);
-
-        // Create a new JsonSerializerOptions for this specific endpoint
-        var options = new JsonSerializerOptions()
-        {
-            ReferenceHandler = ReferenceHandler.IgnoreCycles,
-            WriteIndented = true,
-        };
-
-        // Add the default converters from the global configuration
-        options.Converters.Add(new DateOnlyJsonConverter());
-        options.Converters.Add(new LTreeJsonConverter());
-
-        // Add special converter for the 'Observations' collection
-        options.Converters.Add(new ObservationConverter());
-
-        return new JsonResult(boreholes, options);
-    }
-
-    /// <summary>
-    /// Exports the details of up to <see cref="MaxPageSize"></see> boreholes as a CSV file. Filters the boreholes based on the provided list of IDs.
-    /// </summary>
-    /// <param name="ids">The list of IDs for the boreholes to be exported.</param>
-    /// <returns>A CSV file containing the details specified boreholes.</returns>
-    [HttpGet("export-csv")]
-    [Authorize(Policy = PolicyNames.Viewer)]
-    public async Task<IActionResult> DownloadCsvAsync([FromQuery][MaxLength(MaxPageSize)] IEnumerable<int> ids)
-    {
-        ids = ids.Take(MaxPageSize).ToList();
-        if (!ids.Any()) return BadRequest("The list of IDs must not be empty.");
-
-        var boreholes = await Context.Boreholes
-            .Where(borehole => ids.Contains(borehole.Id))
-            .Select(b => new
-            {
-                b.Id,
-                b.OriginalName,
-                b.ProjectName,
-                b.Name,
-                b.RestrictionId,
-                b.RestrictionUntil,
-                b.NationalInterest,
-                b.LocationX,
-                b.LocationY,
-                b.LocationPrecisionId,
-                b.ElevationZ,
-                b.ElevationPrecisionId,
-                b.ReferenceElevation,
-                b.ReferenceElevationTypeId,
-                b.ReferenceElevationPrecisionId,
-                b.HrsId,
-                b.TypeId,
-                b.PurposeId,
-                b.StatusId,
-                b.Remarks,
-                b.TotalDepth,
-                b.DepthPrecisionId,
-                b.TopBedrockFreshMd,
-                b.TopBedrockWeatheredMd,
-                b.HasGroundwater,
-                b.LithologyTopBedrockId,
-                b.ChronostratigraphyTopBedrockId,
-                b.LithostratigraphyTopBedrockId,
-            })
-            .ToListAsync()
-            .ConfigureAwait(false);
-
-        if (boreholes.Count == 0) return NotFound("No borehole(s) found for the provided id(s).");
-
-        using var stringWriter = new StringWriter();
-        using var csvWriter = new CsvWriter(stringWriter, CultureInfo.InvariantCulture);
-        await csvWriter.WriteRecordsAsync(boreholes).ConfigureAwait(false);
-
-        return File(Encoding.UTF8.GetBytes(stringWriter.ToString()), "text/csv", "boreholes_export.csv");
     }
 
     /// <summary>
@@ -231,7 +140,7 @@ public class BoreholeController : BoreholeControllerBase<Borehole>
             return Unauthorized();
         }
 
-        var borehole = await GetBoreholesWithIncludes()
+        var borehole = await Context.Boreholes.GetAllWithIncludes()
             .AsNoTracking()
             .SingleOrDefaultAsync(b => b.Id == id)
             .ConfigureAwait(false);
@@ -342,34 +251,5 @@ public class BoreholeController : BoreholeControllerBase<Borehole>
     {
         if (entity == null) return default;
         return await Task.FromResult<int?>(entity.Id).ConfigureAwait(false);
-    }
-
-    private IQueryable<Borehole> GetBoreholesWithIncludes()
-    {
-        return Context.Boreholes.Include(b => b.Stratigraphies).ThenInclude(s => s.Layers).ThenInclude(l => l.LayerColorCodes)
-            .Include(b => b.Stratigraphies).ThenInclude(s => s.Layers).ThenInclude(l => l.LayerDebrisCodes)
-            .Include(b => b.Stratigraphies).ThenInclude(s => s.Layers).ThenInclude(l => l.LayerGrainAngularityCodes)
-            .Include(b => b.Stratigraphies).ThenInclude(s => s.Layers).ThenInclude(l => l.LayerGrainShapeCodes)
-            .Include(b => b.Stratigraphies).ThenInclude(s => s.Layers).ThenInclude(l => l.LayerOrganicComponentCodes)
-            .Include(b => b.Stratigraphies).ThenInclude(s => s.Layers).ThenInclude(l => l.LayerUscs3Codes)
-            .Include(b => b.Stratigraphies).ThenInclude(s => s.LithologicalDescriptions)
-            .Include(b => b.Stratigraphies).ThenInclude(s => s.FaciesDescriptions)
-            .Include(b => b.Stratigraphies).ThenInclude(s => s.ChronostratigraphyLayers)
-            .Include(b => b.Stratigraphies).ThenInclude(s => s.LithostratigraphyLayers)
-            .Include(b => b.Completions).ThenInclude(c => c.Casings).ThenInclude(c => c.CasingElements)
-            .Include(b => b.Completions).ThenInclude(c => c.Instrumentations)
-            .Include(b => b.Completions).ThenInclude(c => c.Backfills)
-            .Include(b => b.Sections).ThenInclude(s => s.SectionElements)
-            .Include(b => b.Observations).ThenInclude(o => (o as FieldMeasurement)!.FieldMeasurementResults)
-            .Include(b => b.Observations).ThenInclude(o => (o as Hydrotest)!.HydrotestResults)
-            .Include(b => b.Observations).ThenInclude(o => (o as Hydrotest)!.HydrotestEvaluationMethodCodes)
-            .Include(b => b.Observations).ThenInclude(o => (o as Hydrotest)!.HydrotestFlowDirectionCodes)
-            .Include(b => b.Observations).ThenInclude(o => (o as Hydrotest)!.HydrotestKindCodes)
-            .Include(b => b.BoreholeCodelists)
-            .Include(b => b.Workflows)
-            .Include(b => b.BoreholeFiles)
-            .Include(b => b.BoreholeGeometry)
-            .Include(b => b.Workgroup)
-            .Include(b => b.UpdatedBy);
     }
 }
